[refactor] 페이지 서버 액션 로 분리
