+++ src/app/gatherings/detail/[id]/page.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next'
 import { serverFetcher } from '@/lib/api/serverFetcher';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
+import { getDetailReview } from '@/actions/gathering/detail/getDetailReview';
 import { Gathering } from '@/types/gatherings';
-import { Reviews } from '@/types/reviews';
 import GatheringsDetailUI from '@/components/gatherings/detail/GatheringDetailUI';
 
 export interface PageProps {
@@ -37,43 +37,6 @@ export async function generateMetadata(
         title: `${detail?.name} 상세 정보 | Meet Meet`,
         description: `${detail?.name}의 상세 정보 페이지입니다.`,
     };
-}
-
-/**
- * 모임 상세 페이지의 리뷰 데이터 조회
- * @param id - 모임 ID
- * @returns 모임 리뷰 데이터
- */
-async function getDetailReview(id: string): Promise<Reviews> {
-    try {
-        const data = await serverFetcher<Reviews>(`${EXTERNAL_PATHS.fetchDetailReview(Number(id))}`);
-
-        if (
-            data &&
-            Array.isArray(data.data) &&
-            typeof data.totalItemCount === 'number' &&
-            typeof data.currentPage === 'number' &&
-            typeof data.totalPages === 'number'
-        ) {
-            return data as Reviews;
-        } else {
-            console.warn('응답이 Review 타입이 아닙니다:', data);
-            return {
-                data: [],
-                totalItemCount: 0,
-                currentPage: 0,
-                totalPages: 0,
-            };
-        }
-    } catch (error) {
-        console.error('모임 리뷰 조회 Server Error:', error);
-        return {
-            data: [],
-            totalItemCount: 0,
-            currentPage: 0,
-            totalPages: 0,
-        };
-    }
 }
 
 export default async function GatheringsDetailPage({ params }: PageProps) {
